feat(release): Update plugin version and make it snapshot in post-release (#143)
Incrementing the version before the release sets a fixed version until the next release it's made. For testing purposes, it's interesting to have a different version name that doesn't match the name of the version in production (ending with `-SNAPSHOT`). Versions are incremented by increasing by one the last numeric value; for example, `1.2.3` will become `1.2.4`, and `1.2.5-beta.1` will become `1.2.5-beta.2`. The first version in these examples is the released version (the argument Craft takes). Note, however, that versions must end with a number: `1.2.3-alpha` is a semantic version but not compatible with this post-release script (doesn't end with a number), and `1.2.3-alpha.0` should be used instead.