Update Maven Publish Plugin
This updates the Maven Publish Plugin which should unblock enabling the
configuration cache and also help with updating to Gradle 8.x in this
repo.

The changelog for all the versions is here: https://github.com/vanniktech/gradle-maven-publish-plugin/blob/main/CHANGELOG.md#version-0180-2021-09-13

Note, we cannot update to a higher version until the build itself is
build with Gradle 8.x or higher.

The api to configure signing was changed in 0.20.0 this is why the
gradle property was adjusted.