Remove code for backwards-compatibility testing
This test code was added in #576. It ensures that the Gradle plugin gracefully handles the case where we attempt to upload a ProGuard mapping, with release association, to a self-hosted instance which does not yet support ProGuard release association. Since we no longer will support ProGuard release association in the Gradle plugin, this test code has become irrelevant.